requirements.txt
requests==2.32.4
python-dotenv==1.1.1
openai==1.98.0
pandas==2.3.1
tqdm==4.67.1
tabulate==0.9.0
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-qdrant==0.2.0
qdrant-client==1.15.1
langchain-huggingface==0.3.1
sentence-transformers==5.0.0
transformers==4.54.1
torch==2.7.1
lmstudio==1.4.1
ipywidgets==8.1.7
